% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{aggreg_stratdata_in_harmonclasses}
\alias{aggreg_stratdata_in_harmonclasses}
\title{Aggregating a stratified dataset into the newly added harmonised classes}
\usage{
aggreg_stratdata_in_harmonclasses(
  df,
  harmon_var_col,
  former_var_col,
  marg_dist_collist
)
}
\arguments{
\item{df}{The stratified dataframe which contains old classes and a new harmonised class (corresponding to another dataframe) as well as some marginal distributions}

\item{harmon_var_col}{The name of the column that contains the new harmonised classes (e.g. output of varclass_harmonization function)}

\item{former_var_col}{The name of the old column that contains the classes that were initially in the dataframe but which we want to get rid off}

\item{marg_dist_collist}{A list of names of the columns that specify the marginal distributions (numbers/counts). These will have to be aggregated into the new harmonised classes}
}
\value{
new dataframe with only the harmonised classes, other stratified variables that were in the df, and the aggregated marginal distributions.
}
\description{
Once the new harmonised classes of have been added to a stratified dataframe, the marginal distributions have to be aggregated correspondingly. This function helps doing exactly that, resulting in a new dataframe with only the harmonised classes, other stratified variables that were in the df, and the aggregated marginal distributions.
}
