% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conditional_prob.R
\name{calc_propens_agents}
\alias{calc_propens_agents}
\title{Calculating the conditional propensity to have an attribute based on conditional variables}
\usage{
calc_propens_agents(
  dataframe,
  variable,
  total_population,
  agent_df,
  list_conditional_var
)
}
\arguments{
\item{dataframe}{stratified dataframe which includes all combinations of the conditional variables (the ones based on which we want to calculate the propensity)}

\item{variable}{the number of people with the variable class for which we want to calculate the propensity}

\item{total_population}{the total number of people within the combination of conditional variables}

\item{agent_df}{agent dataset, the full dataset of individual agents that was previously created and that contains the conditional variables}

\item{list_conditional_var}{list of conditional variables, the variables on which basis the propensity to have an attribute should be created}
}
\value{
The output is the agent dataframe with an additional column of their propensity/probability to have the respective property/attribute.
}
\description{
This is a function to calculate the probability/propensity to have a specific class of a variable conditional on other variables. It requires a stratified dataframe (@dataframe) which includes all combinations of the conditional variables (the ones based on which we want to calculate the propensity), the number of people with the variable class (@variable) for which we want to calculate the propensity, and the total number of people within the combination of conditional variables (@total_population). The function calculates the propensity to have a specific class based on the conditional variables, by dividing the number of people with the class if interest by the total number of people within the combination of conditional variables. Further, the function joins these probabilities with the agent dataset (@agent_df) based on the list of conditional variables (@list_conditional_var). This requires the variable/column names of the conditional variables to be equal for the stratified and the agent dataframe. Finally, it shuffles the agent dataset to avoid biases due to a non random order in the next steps.
}
