% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{create_stratified_prob_table}
\alias{create_stratified_prob_table}
\title{Creates a stratified probability table from single attribute propensities}
\usage{
create_stratified_prob_table(
  nested_cond_attr_list,
  column_names,
  orig_df,
  strat_var,
  var_for_pred,
  total_population
)
}
\arguments{
\item{column_names}{the names of columns}

\item{total_population}{}
}
\description{
This function
}
