% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{crosstabular_stratfieddf_to_singleside_df}
\alias{crosstabular_stratfieddf_to_singleside_df}
\title{Crosstabular Stratified Format to Single Sided Variable Combination Format}
\usage{
crosstabular_stratfieddf_to_singleside_df(
  df,
  nrow_var,
  ncol_var,
  row_var_names,
  col_var_names
)
}
\arguments{
\item{df}{The stratified dataset that you want to process}

\item{nrow_var}{number of rows corresponding to variable names}

\item{ncol_var}{number of columns corresponding to variable names}

\item{row_var_names}{The variable names that you want to use to describe the variables in the rows. This will be used as variable column name for the output dataset.The order needs to be from top row to lower row variables.}

\item{col_var_names}{The variable names that you want to use to describe the variables in the columns. This will be used as variable column name for the output dataset. The order needs to be from left column to right column variables.}
}
\value{
Returns a processed stratefied dataframe with all variables as columns on the left. There will be all unique variable combinations. Additionally there will be the counts variable that shows how many people there are for the respective variable combination.
}
\description{
In order to facilitate the data preparation, this function can transform any stratified dataset structure in terms of number of row variables and column variable combinations into a dataset structure of all variable combinations as columns in the left and a single "counts" column on the right.
}
