% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generat_agent_df.R
\name{gen_agent_df}
\alias{gen_agent_df}
\title{Generating an agent dataframe of the population size and assigning a unique ID}
\usage{
gen_agent_df(pop_size)
}
\arguments{
\item{pop_size}{The size of the population for the are for which a synthetic agent population should be created.}
}
\value{
Returns a new dataframe of len(pop_size) with unique agent_ids.
}
\description{
The function creates a new agent dataframe of an indicated population size to start the building of the synthetic agent population.  Each row in the dataframe represents an individual and is assigned a unique ID. This is basically the agent dataframe scelaton that will be iteratively extended with attribute data based on conditional and marginal distributions of demographics.
}
