% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_agent_values.R
\name{get_agent_values_from_fractions}
\alias{get_agent_values_from_fractions}
\title{Generate Agent Attribute Values Based on Group Fractions}
\usage{
get_agent_values_from_fractions(
  group_fractions,
  group_agent_count,
  target_attribute = "target_attribute"
)
}
\arguments{
\item{group_fractions}{A data frame containing the fractions for each value of the target attribute.
The data frame must include a \code{fraction} column representing the proportion of each target attribute value within the group,
and a column for the target attribute itself.}

\item{group_agent_count}{An integer representing the total number of agents in the group.}

\item{target_attribute}{A string specifying the name of the target attribute column in \code{group_fractions}.
Default is "target_attribute".}
}
\value{
A vector of the target attribute values for all agents in the group, with the values distributed according to the input fractions.
}
\description{
This function assigns values of a target attribute to agents within a group based on the relative proportions
(fractions) of those values in the group. The generated values match the proportions given by the input fractions.
}
\details{
This function takes a group of agents and assigns values to the target attribute based on the provided \code{group_fractions}.
The function works by:
\enumerate{
\item \strong{Calculating group counts}: The number of agents corresponding to each fraction is computed by multiplying
the total number of agents in the group (\code{group_agent_count}) by each fraction.
\item \strong{Generating values}: The function replicates each target attribute value according to its count and then
randomly shuffles these values to avoid any ordering bias.
}
}
\examples{
# Example group fractions
group_fractions <- data.frame(
  target_attribute = c("young", "middle-aged", "senior"),
  fraction = c(0.2, 0.5, 0.3)
)

# Generate values for a group of 10 agents
agent_values <- get_agent_values_from_fractions(group_fractions, 
                                               group_agent_count = 10, 
                                               target_attribute = "target_attribute")

}
