% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_proportionate_fitting.R
\name{suppress_probability_warning}
\alias{suppress_probability_warning}
\title{Suppress Specific Probability Warning}
\usage{
suppress_probability_warning(expr)
}
\arguments{
\item{expr}{An R expression to be evaluated. It can be any function or block of code that
potentially generates warnings.}
}
\value{
The result of the evaluated expression. Warnings other than the specified one are
displayed, and the execution continues as normal.
}
\description{
This function evaluates an expression while suppressing a specific warning message:
"Target not consistents - shifting to probabilities!". All other warnings are displayed
as usual. This is useful when using functions like \code{Ipfp} where such warnings may occur,
but you want to ignore them without suppressing other warnings.
}
\details{
The function uses \code{withCallingHandlers} to handle warnings during the evaluation
of the expression. It checks if the warning message contains the text
"Target not consistents - shifting to probabilities!" and suppresses it. Other warnings
are not affected.
}
