% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{create_stratified_prob_table}
\alias{create_stratified_prob_table}
\title{Creates a stratified probability table from single attribute propensities}
\usage{
create_stratified_prob_table(
  df,
  nested_cond_attr_list,
  cond_var_names,
  cond_attr_column,
  var_for_pred,
  total_population
)
}
\arguments{
\item{df}{A dataframe with one column in which all classes of the conditional variables are listed in seperate rows and additional columns that show the marginal distributions for a variable of interest}

\item{nested_cond_attr_list}{a nested list of classes of conditional variables. (e.g. list(c("female", "male", "non-binary"), c("employed", "unemployed")))}

\item{cond_var_names}{a list of the names that you want to give the variables to which the  classes indicated in the nested_cond_attr_list belong. It has to have the same order as that list. (e.g. for above that could be c("sex", "employ_status"))}

\item{cond_attr_column}{The column in which all separate conditional variable classes are listed.}

\item{var_for_pred}{The columnnames that indicate the classes of the variable for which the stratified dataframe should be generated, if it is a binary variable, then only one class needs to be given, since the propensity to have the other class is class is 1-the propensity of the indicated class.}

\item{total_population}{The column indicating the total population that has the conditional variable attribute of the row}
}
\value{
a stratified dataset with all combinations of the conditional variables and the propensity to be/have the indicated classes
}
\description{
This function combines multiple - joint distributions of conditional variables into one stratified dataframe of all-variable combinations and the propensity to have the variable classes of interest
}
\examples{
## creating mock data
cond_variables = c("male", "female", "non-binary", "single_household", "multi_pers_household")
employed = sample(40:400,length(cond_variables))
unemployed = sample(40:400,length(cond_variables))
df = data.frame(cond_variables, employed , unemployed)
df$tot_pop = df$employed + df$unemployed

# function application
strat_df = create_stratified_prob_table(df = df, nested_cond_attr_list = list(c("male", "female", "non-binary"), c("single_household", "multi_pers_household")), cond_var_names = c("sex", "hh_structure"), cond_attr_column = "cond_variables", var_for_pred = c("employed", "unemployed"), total_population = "tot_pop" )
print(strat_df)
}
