% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_fractions.R
\name{get_group_fractions}
\alias{get_group_fractions}
\title{Calculate Group Fractions from Contingency Data}
\usage{
get_group_fractions(
  df_contingency,
  group_by,
  target_attribute,
  margins_names = NULL,
  margins_group = NULL
)
}
\arguments{
\item{df_contingency}{A data frame representing the contingency table that contains the observed distribution of agents across combinations of attributes.}

\item{group_by}{A character vector specifying the columns in the contingency table to group the data by. This typically represents the dimensions or attributes along which the data is split.}

\item{target_attribute}{A string specifying the target attribute for which the fractions are calculated. This is the variable of interest in the contingency table.}

\item{margins_names}{A list of column names representing the margins (optional). If provided, it includes additional attributes to consider when calculating the fractions.}

\item{margins_group}{A character vector indicating the group for margins (optional). This allows for finer resolution in calculating fractions based on specific margin groups.}
}
\value{
A data frame of calculated fractions, with rows corresponding to combinations of the target attribute
and any provided margins, and columns representing the fractions.
}
\description{
This function computes the fractions (proportions) of synthetic agents for each combination of group attributes
based on a contingency table. The resulting fractions represent the distribution of the target attribute across
specified groups and optional margins.
}
\details{
The function first calls \code{calculate_fractions} to generate fractional values from the contingency table.
These fractions represent the proportion of agents assigned to each combination of the target attribute and the
group-by variables. The result is a filtered and ordered data frame of fractions, with indices set by the target
attribute and any margin names, if applicable.

If \code{margins_names} are provided, the function combines them with the \code{target_attribute} and any additional
margin groups to ensure the returned fractions are indexed by all relevant attributes.
}
\examples{
# Example: Calculate fractions for target attribute 'age_group' 
# across 'sex' and 'migrationbackground' groups
df_contingency <- data.frame(
  sex = c("male", "female", "male", "female"),
  migrationbackground = c("Dutch", "Dutch", "Non-Dutch", "Non-Dutch"),
  age_group = c("0-15", "15-25", "25-45", "45-65"),
  count = c(30, 25, 50, 45)
)

fractions <- get_group_fractions(df_contingency, group_by = c("sex", "migrationbackground"), 
                                 target_attribute = "age_group")
# Returns a data frame of fractions based on 'age_group', 'sex', and 'migrationbackground'

}
\keyword{internal}
