% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{restructure_one_var_marginal}
\alias{restructure_one_var_marginal}
\title{Restructures a single-sided stratified dataframe so that the classes of one column/variable of interest are seperate columns}
\usage{
restructure_one_var_marginal(df, variable, countsname)
}
\arguments{
\item{df}{The original stratified dataframe with all varable combinations as columns on the left side.}

\item{variable}{The variable of interest, for which on wants to generate seperate columns of the subclasses.}

\item{countsname}{The columnname of the column in the original datafram, which indicates the counts for all variable combinations.}
}
\value{
the output will be a dataframe of all variable combinations excluding the variable of interest and the marginal distributions of the variable combinations along the classes of the variable of interest.
}
\description{
This function takes a single-sided stratified dataframe, such as the output of the crosstabular_to_singleside_df function and restructures it as such that the unique classes of one variable of interest (any of the variablecolumns), will become seperate columns. Hence the output will be a dataframe of all variable combinations excluding the variable of interest and the marginal distributions of the variable combinations along the classes of the variable of interest.
}
\section{Example Tables (random numbers)}{
\subsection{Turns a single side stratified dataframe such as this:}{\tabular{cccc}{
   age_group \tab sex \tab employ_status \tab counts \cr
   A1 \tab male \tab employed \tab 25 \cr
   A2 \tab male \tab employed \tab 132 \cr
   A3 \tab male \tab employed \tab 122 \cr
   A1 \tab male \tab unemployed \tab 133 \cr
   A2 \tab male \tab unemployed \tab 323 \cr
   A3 \tab male \tab unemployed \tab 360 \cr
   A1 \tab female \tab employed \tab 175 \cr
   A2 \tab female \tab employed \tab 275 \cr
   A3 \tab female \tab employed \tab 394 \cr
   A1 \tab female \tab unemployed \tab 389 \cr
   A2 \tab female \tab unemployed \tab 206 \cr
   A3 \tab female \tab unemployed \tab 25 \cr
   A1 \tab non-binary \tab employed \tab 196 \cr
   A2 \tab non-binary \tab employed \tab 248 \cr
   A3 \tab non-binary \tab employed \tab 137 \cr
   A1 \tab non-binary \tab unemployed \tab 203 \cr
   A2 \tab non-binary \tab unemployed \tab 270 \cr
   A3 \tab non-binary \tab unemployed \tab 34 \cr
}

}

\subsection{into a one variable marginal distribution table for a specified variable of interest, and adds a total for the population of the attribute combination}{\tabular{ccccc}{
   age_group \tab sex \tab employed \tab un-employed \tab total \cr
   A1 \tab male \tab 25 \tab 133 \tab 158 \cr
   A2 \tab male \tab 132 \tab 323 \tab 455 \cr
   A3 \tab male \tab 122 \tab 360 \tab 482 \cr
   A1 \tab female \tab 175 \tab 389 \tab 564 \cr
   A2 \tab female \tab 275 \tab 206 \tab 481 \cr
   A3 \tab female \tab 394 \tab 25 \tab 419 \cr
   A1 \tab non-binary \tab 196 \tab 203 \tab 399 \cr
   A2 \tab non-binary \tab 248 \tab 270 \tab 518 \cr
   A3 \tab non-binary \tab 137 \tab 34 \tab 171 \cr
}

}
}

\examples{
## generating some example mock data ##
# stratified dataframe mock data, can be output of function: crosstabular_to_singleside_df
age_group = rep(c("A1", "A2", "A3", "A4"), 6)
sex = rep(c("male", "female", "non-binary"), each = 8)
employ_status = rep(rep(c("employed", "unemployed"), each = 4),3)
counts = sample(1:400,length(age_group))
singleside_stratified_df = data.frame(age_group, sex , employ_status, counts)

# function application
one_var_marginal_df = restructure_one_var_marginal(singleside_stratified_df, "employ_status", "counts")
print(one_var_marginal_df)

}
