% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_attribute.R
\name{verify_target_attribute}
\alias{verify_target_attribute}
\title{Verify the Target Attribute}
\usage{
verify_target_attribute(df, df_contingency, target_attribute, margins_group)
}
\arguments{
\item{df}{A data frame representing the synthetic population that includes the target attribute.}

\item{df_contingency}{A data frame representing the original distribution from which the target attribute
is derived.}

\item{target_attribute}{A string representing the name of the target attribute to verify.}

\item{margins_group}{A vector of attribute names used to group the data for comparison.}
}
\value{
NULL This function does not return any value; it generates warnings if there are issues
with the target attribute.
}
\description{
This function verifies the integrity of the target attribute in the synthetic population.
It checks if any agents have not been assigned a value for the target attribute and compares
the distribution of the target attribute with a given contingency table using a Z-squared test.
}
